requirements.txt
amqp==5.0.7
asgiref==3.4.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==21.12b0
cachetools==4.2.4
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0
django-filter==21.1
djangorestframework==3.13.1
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
kombu==5.2.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
prompt-toolkit==3.0.24
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
pytz==2021.3
redis==4.0.2
requests==2.26.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
